Bumping max translucent quads from short to int

+++ src/main/java/me/cortex/nvidium/sodiumCompat/RepackagedSectionOutput.java
@@ -6,7 +6,7 @@ import org.joml.Vector3i;
 //Computed on the build thread instead of the render thread saving alot of 1% lows
 public record RepackagedSectionOutput(int quads,
                                       NativeBuffer geometry,
-                                      short[] offsets,
+                                      int[] offsets,
                                       Vector3i min,
                                       Vector3i size) {
     public void delete() {
